+++ src/commands/index.ts
@@ -1,3 +1,6 @@
 export { sendMsg } from "./send.js";
 export { kick } from "./kick.js";
 export { ban } from "./ban.js";
+export { rmFromGroup } from "./rmfromgroup.js";
+export { archive } from "./archive.js";
+export { fetchMsgs } from "./fetch.js";